misc: addressed review comments

+++ backend/src/db/migrations/20240702055253_add-encrypted-webhook-url.ts
@@ -5,45 +5,49 @@ import { WebhookType } from "@app/services/webhook/webhook-types";
 import { TableName } from "../schemas";
 
 export async function up(knex: Knex): Promise<void> {
-  const hasEncryptedURL = await knex.schema.hasColumn(TableName.Webhook, "encryptedUrl");
+  const hasUrlCipherText = await knex.schema.hasColumn(TableName.Webhook, "urlCipherText");
   const hasUrlIV = await knex.schema.hasColumn(TableName.Webhook, "urlIV");
   const hasUrlTag = await knex.schema.hasColumn(TableName.Webhook, "urlTag");
   const hasType = await knex.schema.hasColumn(TableName.Webhook, "type");
 
-  await knex.schema.alterTable(TableName.Webhook, (tb) => {
-    if (!hasEncryptedURL) {
-      tb.text("encryptedUrl");
-    }
-    if (!hasUrlIV) {
-      tb.string("urlIV");
-    }
-    if (!hasUrlTag) {
-      tb.string("urlTag");
-    }
-    if (!hasType) {
-      tb.string("type").defaultTo(WebhookType.GENERAL);
-    }
-  });
+  if (await knex.schema.hasTable(TableName.Webhook)) {
+    await knex.schema.alterTable(TableName.Webhook, (tb) => {
+      if (!hasUrlCipherText) {
+        tb.text("urlCipherText");
+      }
+      if (!hasUrlIV) {
+        tb.string("urlIV");
+      }
+      if (!hasUrlTag) {
+        tb.string("urlTag");
+      }
+      if (!hasType) {
+        tb.string("type").defaultTo(WebhookType.GENERAL);
+      }
+    });
+  }
 }
 
 export async function down(knex: Knex): Promise<void> {
-  const hasEncryptedURL = await knex.schema.hasColumn(TableName.Webhook, "encryptedUrl");
+  const hasUrlCipherText = await knex.schema.hasColumn(TableName.Webhook, "urlCipherText");
   const hasUrlIV = await knex.schema.hasColumn(TableName.Webhook, "urlIV");
   const hasUrlTag = await knex.schema.hasColumn(TableName.Webhook, "urlTag");
   const hasType = await knex.schema.hasColumn(TableName.Webhook, "type");
 
-  await knex.schema.alterTable(TableName.Webhook, (t) => {
-    if (hasEncryptedURL) {
-      t.dropColumn("encryptedUrl");
-    }
-    if (hasUrlIV) {
-      t.dropColumn("urlIV");
-    }
-    if (hasUrlTag) {
-      t.dropColumn("urlTag");
-    }
-    if (hasType) {
-      t.dropColumn("type");
-    }
-  });
+  if (await knex.schema.hasTable(TableName.Webhook)) {
+    await knex.schema.alterTable(TableName.Webhook, (t) => {
+      if (hasUrlCipherText) {
+        t.dropColumn("urlCipherText");
+      }
+      if (hasUrlIV) {
+        t.dropColumn("urlIV");
+      }
+      if (hasUrlTag) {
+        t.dropColumn("urlTag");
+      }
+      if (hasType) {
+        t.dropColumn("type");
+      }
+    });
+  }
 }
